Split the current weather component into two seperate files so now it has a CSS file and JSX file, updated some CSS for the rating page, Created and finished the README file, added an visibility image to the assets folder, added it to the current weather page
