+++ src/App.css
@@ -3,7 +3,8 @@
 .App {
   font-family: 'Bebas Neue', cursive;
   background-color: #1f2987;
-  height: 100vh;
+  min-height: 100vh;
+  height: auto;
 }
 
 .nav-bar ul {
@@ -96,6 +97,7 @@ color : #2f3640;
   text-align: center;
   padding-top: 5%;
 }
+
 .about-section h1 {
   margin: 0px;
 }
@@ -197,6 +199,7 @@ color : #2f3640;
 .rate-btn {
   display: flex;
   justify-content: center;
+  margin-bottom: 5%;
 }
 
 .rate-btn button {
@@ -224,131 +227,3 @@ color : #2f3640;
   transform: translateY(-1px);
 }
 
-.current-weather-main-container {
-  display: flex;
-  justify-content: center;
-}
-
-.current-weather-des-container {
-  background-color: #2F3740;
-  color: white;
-  width: 80%;
-  height: auto;
-  margin-top: 2%;
-  border-radius: 10%;
-}
-.current-weather-grid-container {
-  display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  grid-template-rows: auto auto auto auto;
-}
-.current-weather-grid-container p {
-  margin: 0px;
-}
-.current-weather-grid-container h1 {
-  grid-column: span 12;
-  text-align: center;
-}
-
-.current-weather-temp {
-  grid-column-start: 3;
-  grid-column-end: 6;
-  grid-row-start: 2;
-  grid-row-end: 3;
-  margin-bottom: 8%;
-}
-.current-weather-temp p:nth-child(1) {
-  font-size: 2em;
-}
-.current-weather-temp p:nth-child(2) {
-  font-size: 0.7em;
-}
-.current-weather-temp p:nth-child(3) {
-  font-size: 0.7em;
-  margin: 8% 0% 0% 0%;
-}
-.current-weather-temp p:nth-child(4) {
-  font-size: 0.7em;
-  margin: 0% 0% 0% 0%;
-}
-
-.current-weather-icon-des {
-  grid-column-start: 9;
-  grid-column-end: 12;
-  grid-row-start: 2;
-  grid-row-end: 3;
-}
-
-.current-weather-des {
-  margin-left: 10%;
-}
-
-.current-weather-feels-like {
-  font-size: 0.6em;
-}
-
-.current-weather-pop {
-  grid-column-start: 3;
-  grid-column-end: 11;
-  grid-row-start: 3;
-  grid-row-end: 4;
-  margin-bottom: 5%;
-}
-.current-weather-pop img {
-  width: 40px;
-  height: 40px;
-}
-.current-weather-vis {
-  grid-column-start: 9;
-  grid-column-end: 12;
-  grid-row-start: 3;
-  grid-row-end: 4;
-}
-
-/* .weather-btm-container {
-  grid-column-start: 3;
-  grid-column-end: 11;
-  grid-row-start: 4;
-  grid-row-end: 5;
-  display: flex;
-  justify-content: space-between;
-} */
-
-.current-weather-wind {
-  grid-column-start: 3;
-  grid-column-end: 5;
-  grid-row-start: 4;
-  grid-row-end: 5;
-  margin-bottom: 5%;
-}
-.current-weather-wind img {
-  width: 40px;
-  height: 40px;
-}
-
-.current-weather-hum {
-  grid-column-start: 6;
-  grid-column-end: 8;
-  grid-row-start: 4;
-  grid-row-end: 5;
-}
-.current-weather-hum img {
-  width: 40px;
-  height: 40px;
-}
-.current-weather-uvi {
-  grid-column-start: 9;
-  grid-column-end: 11;
-  grid-row-start: 4;
-  grid-row-end: 5;
-}
-.current-weather-uvi img {
-  width: 40px;
-  height: 40px;
-}
-
-.crt-weather-load {
-  grid-column-start: 3;
-  grid-column-end: 10;
-  padding-bottom: 10%;
-}
