+++ src/components/CurrentRating.jsx
@@ -1,13 +1,16 @@
 //imports the useContext hook from react
 import React, { useContext } from "react";
 import { MyContext } from "../context/MyProvider.js";
-
 import ProgressBar from "./StatusBar/ProgressBar";
+
 
 //changed to curly brackets, because there will be a return
 const CurrentRating = () => {
   //the varible is equal to the useContext hook with the argument of MyContext, this will give us our data instead of all the everthing.
   const context = useContext(MyContext);
+
+  // let circleColor = "green";
+
 
   const state = {
     size: 180,
@@ -29,19 +32,19 @@ const CurrentRating = () => {
               />
             </div>
             <div className="rating-params">
-              <p>The parameters we checked for your day were:</p>
+              <p>The weather factors are:</p>
               <p>
-                1. The real feel temperature is:{" "}
-                {context.weather.current.feels_like}°C
+                1. Real feel temperature is:{"   "}
+                {context.weather.current.feels_like.toFixed(0)}°C
               </p>
               <p>
-                2. The wind speed is: {context.crtWindSpeed}km/h
+                2. Wind speed is: {context.crtWindSpeed}km/h
               </p>
               <p>
-                3. The rain probability: {context.weather.hourly[0].pop * 100}%
+                3. Rain probability: {(context.weather.hourly[0].pop * 100).toFixed(0)}%
               </p>
-              <p>4. The UV index is: {context.weather.current.uvi}</p>
-              <p>5. The air quality is: {context.airPollutionDes}</p>
+              <p>4. UV index is: {context.weather.current.uvi}</p>
+              <p>5. Air quality is: {context.airPollutionDes}</p>
             </div>
             <div className="rate-btn">
               <button onClick={context.handleNavCurrentWeather}>
