+++ src/components/CurrentWeather/CurrentWeather.css
@@ -1,0 +1,139 @@
+
+.current-weather-main-container {
+    display: flex;
+    justify-content: center;
+  }
+  
+  .current-weather-des-container {
+    background-color: #2F3740;
+    color: white;
+    width: 80%;
+    height: auto;
+    margin-top: 2%;
+    border-radius: 10%;
+  }
+  .current-weather-grid-container {
+    display: grid;
+    grid-template-columns: repeat(12, 1fr);
+    grid-template-rows: repeat(4, auto);
+  }
+  .current-weather-grid-container p {
+    margin: 0px;
+  }
+  .current-weather-grid-container h1 {
+    grid-column: span 12;
+    text-align: center;
+  }
+  
+  .current-weather-temp {
+    grid-column-start: 3;
+    grid-column-end: 7;
+    grid-row-start: 2;
+    grid-row-end: 3;
+    margin-bottom: 8%;
+  }
+  .current-weather-temp p:nth-child(1) {
+    font-size: 2em;
+  }
+  .current-weather-temp p:nth-child(2) {
+    font-size: 0.7em;
+  }
+  .current-weather-temp p:nth-child(3) {
+    font-size: 0.7em;
+    margin: 8% 0% 0% 0%;
+  }
+  .current-weather-temp p:nth-child(4) {
+    font-size: 0.7em;
+    margin: 0% 0% 0% 0%;
+  }
+  
+  .current-weather-icon-des {
+    grid-column-start: 9;
+    grid-column-end: 11;
+    grid-row-start: 2;
+    grid-row-end: 3;
+    text-align: center;
+  }
+  
+  .current-weather-des {
+    margin-left: 10%;
+  }
+  
+  .current-weather-feels-like {
+    font-size: 0.6em;
+  }
+  
+  .current-weather-pop {
+    grid-column-start: 3;
+    grid-column-end: 5;
+    grid-row-start: 3;
+    grid-row-end: 4;
+    margin-bottom: 5%;
+    text-align: center;
+  }
+  .current-weather-pop p {
+    font-size: 1em;
+  }
+
+  .current-weather-pop img {
+    width: 40px;
+    height: 40px;
+  }
+
+  .current-weather-vis {
+    grid-column-start: 9;
+    grid-column-end: 11;
+    grid-row-start: 3;
+    grid-row-end: 4;
+    text-align: center;
+  }
+
+  .current-weather-vis img {
+    width: 40px;
+    height: 40px;
+  }
+  
+  .current-weather-wind {
+    grid-column-start: 2;
+    grid-column-end: 6;
+    grid-row-start: 4;
+    grid-row-end: 5;
+    margin-bottom: 5%;
+    text-align: center;
+    font-size: 1em;
+  }
+  .current-weather-wind img {
+    width: 40px;
+    height: 40px;
+  }
+  
+  .current-weather-hum {
+    grid-column-start: 6;
+    grid-column-end: 8;
+    grid-row-start: 4;
+    grid-row-end: 5;
+    text-align: center;
+  }
+
+  .current-weather-hum img {
+    width: 40px;
+    height: 40px;
+  }
+  .current-weather-uvi {
+    grid-column-start: 9;
+    grid-column-end: 11;
+    grid-row-start: 4;
+    grid-row-end: 5;
+    text-align: center;
+    font-size: 1em;
+  }
+  .current-weather-uvi img {
+    width: 40px;
+    height: 40px;
+  }
+  
+  .crt-weather-load {
+    grid-column-start: 3;
+    grid-column-end: 10;
+    padding-bottom: 10%;
+  }
